+++ R/functions-TopDownSet.R
@@ -79,6 +79,11 @@
 #' column names of the [colData()]
 #' used to define a sample (technical replicate). This is used to add the
 #' `Sample` column (used for easier aggregation, etc.).
+#' @param conditions `character`/`numeric`, one of:
+#'  - `"FilterString"` (default): create condition IDs based on mass labels in
+#'    the *FilterString* column (see *massLabeling* in [writeMethodXmls()] for
+#'    details).
+#'  - A single `numeric` value giving the number of conditions.
 #' @param verbose `logical`, verbose output?
 #' @return A `TopDownSet` object.
 #' @export
@@ -117,6 +122,7 @@ readTopDownFiles <- function(path, pattern=".*",
                                              "CidActivation",
                                              "HcdActivation",
                                              "UvpdActivation"),
+                             conditions="FilterString",
                              verbose=interactive()) {
 
 
@@ -137,11 +143,18 @@ readTopDownFiles <- function(path, pattern=".*",
         verbose=verbose
     )
 
-    scanConditions <-
-        .rbind(lapply(files$txt, .readScanHeadsTable, verbose=verbose))
+    scanConditions <- .rbind(
+        lapply(
+            files$txt,
+            .readScanHeadsTable,
+            conditions=conditions,
+            verbose=verbose
+        )
+    )
 
-    headerInformation <-
-        .rbind(lapply(files$csv, .readExperimentCsv, verbose=verbose))
+    headerInformation <- .rbind(
+        lapply(files$csv, .readExperimentCsv, verbose=verbose)
+    )
 
     mzml <- mapply(
         .readMzMl,